Refactor Model into class
I did this as part of an overarching plan to consolidate each part of
the MVC pattern into a single entry point. By refactoring the Model into
a class, I can have the Controller create and store it internally
without leaving it open and available for other parts of code to
interact with it directly.

The next step will be to refactor the Controller as a class, which will
allow it to be stored in any arbitrary location, whether that be the
global `window` object or elsewhere. From there, I will create a general
View class (which will be much simpler than the current view.ts) which
will be able to be used to add Views quickly and easily to the
Controller.