Separate PersonDefinition from Person
I did this so that I could define and object that holds the information
of a person without it actually being an instantiated person.
Specifically, it was so that I didn't have to manually assign IDs to
each default Person, as that has the potential to lead to a state that
shouldn't be possible. (i.e. multiple people with the same ID, or a
negative ID.)