Change generateNextId() to not track variable
This change makes it so that nextId doesn't need to be tracked, which
reduces the complication of the model. In some cases this would make
sense, but for this app the only requirement is that no ID repeats
within the current model. However, if a Person is deleted and replaced
by another Person with the same ID, this is not an issue.