Fix removePerson method
It appears my original intention was to have it remove a person from the
list of default persons. However, in the context of my refactoring, I
want modification of the defaults to be done separately. This function
now removes a Person from the active list of Persons.

There was also a bug where it was removing a person based on their index
and not on their ID. This caused incorrect behavior when removing
multiple Persons, because after the first one, their IDs no longer sync
with their index in the array.