No longer sort by crossTime after adding Person
I assume that it did this because the model calculations require the
Persons to be sorted in order of crossTime. However, I feel that it
would be better to allow an arbitrary order of Persons, and for the
model to sort them at the point of calculation when that order becomes
necessary. Otherwise, we're mixing the calculation process with the
addPerson() method, which are two separate logic centers and should be
uncoupled from one another.