Throw ValueError when bridge width too small
Since I'm creating custom errors and wanted to check a value inside the
error object, I had to use something other than toThrow() in the test.

I'll probably need to do the same for other error-throwing tests in the
future when I need to check for a particular value in the error object.