Change setPersonSide to find Person directly
That is to say, it doesn't use the getPersonById method like it used to.
The reason for this change is because I'm in the process of modifying
the getPersonById and getPersonByName methods to return copies of the
Person who is found rather than the original object. However, this
change interferes with the functionality of this method.