Create Preset class
This class allows us to define starting configurations to easily save
and switch between them. I decided to implement it now because I
realized that the current way of doing it with a "default" set of data
was flawed, since it wasn't clear whether adding a new Person added it
to the default or to the active model. In other words, if you added a
person, then reset the model, should it keep the person, or return to
how the model was before the person was added? Creating presets helps to
solve this issue by allowing the model to synchronize its active
attributes to one or more Presets.