Add clone method
Note that I made sure to create a copy of the Person objects as well.
This shouldn't really matter right now, since there's no way to directly
modify an existing Preset; the expectation is that if you want to modify
a Preset you just create a new one with modified data. However, in the
event that that changes someday, I figure it's safer to go ahead and
create copies of the Persons anyway so that it doesn't bite me in the
trasero down the road.