Add validation to the Preset constructor
Specifically, the name cannot be an empty string, and the bridge width
cannot be less than 2.

I should probably create a constants file at some point to put things
like the minimum bridge width and reference it instead of hardcoding the
2, but in this case it's a fundamental logic thing, rather than an
arbitrary number. It would still be a good idea, though.