Add import checks for missing or invalid name
There's probably a cleaner way to catch these areas, perhaps with a JSON
schema. For now, given the lack of complexity of the Presets, this is
the fastest way. However, if they start to grow in complexity, it would
be worthwhile to develop an alternative method for defining the
validation.