Add check to make sure Persons are objects
Note that I also had to make a change to the ObjectError class.
Previously, it required the subject of the error to be of type object.
The assumption was that if it wasn't an object, it would instead use
ValueError.

However, I realized that in this case, the point of it being an
ObjectError was not necessarily because the problematic part is an
object, but rather than there was an error with the object as a whole,
or a component part of a larger object: in this case, one of the Persons
within the larger `people` array inside of the Preset object.

I therefore changed the type definition to be

{[id: string]: any} & {object: any}

instead of

{[id: string]: any} & {object: object}