Move Model to private property of Controller
The main effect that this has is that the broadcast() method no longer
needs to receive the Model as an argument, and instead uses the
internally stored Model. While this makes the controller technically
more limited, since it can only act on a single Model at a time, in
reality this is much cleaner and provides more understandable and
maintainable code. If multiple models are in play for some reason, then
multiple controllers will need to be configured, or have have a single
controller swap between the models.